ai_server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
supabase==2.9.1
python-dotenv==1.0.1
httpx==0.27.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
slowapi==0.1.9
numpy==1.26.4
openai==1.58.1
pytest==8.3.3
pytest-asyncio==0.24.0
# Real ML / statistics dependencies (Phase 1 + 2)
pandas==2.2.3
scikit-learn==1.5.2
statsmodels==0.14.4
mlxtend==0.23.1
